Move search components and hooks to features
