+++ src/apps/stable/features/search/components/SearchFields.tsx
@@ -1,11 +1,11 @@
 import React, { type ChangeEvent, type FC, useCallback, useRef } from 'react';
-import AlphaPicker from '../alphaPicker/AlphaPickerComponent';
+import AlphaPicker from '../../../../../components/alphaPicker/AlphaPickerComponent';
 import Input from 'elements/emby-input/Input';
-import globalize from '../../lib/globalize';
-import layoutManager from '../layoutManager';
-import browser from '../../scripts/browser';
+import globalize from '../../../../../lib/globalize';
+import layoutManager from '../../../../../components/layoutManager';
+import browser from '../../../../../scripts/browser';
 import 'material-design-icons-iconfont';
-import '../../styles/flexstyles.scss';
+import 'styles/flexstyles.scss';
 import './searchfields.scss';
 
 interface SearchFieldsProps {
