+++ src/apps/stable/features/search/components/searchfields.scss
