--- src/hooks/searchHook/index.ts
@@ -1,2 +1,0 @@
-export * from './useSearchItems';
-export * from './useSearchSuggestions';
